Allow more than 255 items on single tile

+++ mtanksl.OpenTibia.Common/Objects/Inventory.cs
@@ -25,14 +25,14 @@ namespace OpenTibia.Common.Objects
 
         /// <exception cref="NotSupportedException"></exception>
 
-        public byte AddContent(IContent content)
+        public int AddContent(IContent content)
         {
             throw new NotSupportedException();
         }
 
         /// <exception cref="ArgumentException"></exception>
         
-        public void AddContent(IContent content, byte index)
+        public void AddContent(IContent content, int index)
         {
             if ( !(content is Item) )
             {
@@ -46,7 +46,7 @@ namespace OpenTibia.Common.Objects
 
         /// <exception cref="ArgumentException"></exception>
 
-        public void ReplaceContent(byte index, IContent content)
+        public void ReplaceContent(int index, IContent content)
         {
             if ( !(content is Item) )
             {
@@ -62,7 +62,7 @@ namespace OpenTibia.Common.Objects
             content.Parent = this;
         }
 
-        public void RemoveContent(byte index)
+        public void RemoveContent(int index)
         {
             IContent content = GetContent(index);
 
@@ -73,9 +73,9 @@ namespace OpenTibia.Common.Objects
 
         /// <exception cref="InvalidOperationException"></exception>
 
-        public byte GetIndex(IContent content)
+        public int GetIndex(IContent content)
         {
-            for (byte index = 0; index < contents.Length; index++)
+            for (int index = 0; index < contents.Length; index++)
             {
                 if (contents[index] == content)
                 {
@@ -86,9 +86,9 @@ namespace OpenTibia.Common.Objects
             throw new InvalidOperationException("Content not found.");
         }
 
-        public bool TryGetIndex(IContent content, out byte _index)
+        public bool TryGetIndex(IContent content, out int _index)
         {
-            for (byte index = 0; index < contents.Length; index++)
+            for (int index = 0; index < contents.Length; index++)
             {
                 if (contents[index] == content)
                 {
@@ -103,7 +103,7 @@ namespace OpenTibia.Common.Objects
             return false;
         }
 
-        public IContent GetContent(byte index)
+        public IContent GetContent(int index)
         {
             if (index < 0 || index > contents.Length - 1)
             {
@@ -124,13 +124,13 @@ namespace OpenTibia.Common.Objects
             }
         }
 
-        public IEnumerable< KeyValuePair<byte, IContent> > GetIndexedContents()
+        public IEnumerable< KeyValuePair<int, IContent> > GetIndexedContents()
         {
-            for (byte index = 0; index < contents.Length; index++)
+            for (int index = 0; index < contents.Length; index++)
             {
                 if (contents[index] != null)
                 {
-                    yield return new KeyValuePair<byte, IContent>( index, contents[index] );
+                    yield return new KeyValuePair<int, IContent>( index, contents[index] );
                 }
             }
         }
