+++ mtanksl.OpenTibia.Game/CommandHandlers/PlayerMoveItem/MoveItemWalkToSourceHandler.cs
@@ -13,7 +13,7 @@ namespace OpenTibia.Game.CommandHandlers
             {
                 IContainer beforeContainer = command.Item.Parent;
 
-                byte beforeIndex = beforeContainer.GetIndex(command.Item);
+                int beforeIndex = beforeContainer.GetIndex(command.Item);
 
                 byte beforeCount = command.Count;
 
@@ -30,7 +30,7 @@ namespace OpenTibia.Game.CommandHandlers
                         return Promise.Break;
                     }
 
-                    byte afterIndex = afterContainer.GetIndex(command.Item);
+                    int afterIndex = afterContainer.GetIndex(command.Item);
 
                     byte afterCount = command.Count;
