+++ mtanksl.OpenTibia.Game/CommandHandlers/PlayerUseItemWithCreature/UseItemWithCreatureWalkToTargetHandler.cs
@@ -47,7 +47,7 @@ namespace OpenTibia.Game.CommandHandlers
                 {
                     IContainer beforeContainer = command.Item.Parent;
 
-                    byte beforeIndex = beforeContainer.GetIndex(command.Item);
+                    int beforeIndex = beforeContainer.GetIndex(command.Item);
 
                     return Context.AddCommand(new ParseWalkToUnknownPathCommand(command.Player, command.ToCreature.Tile) ).Then( () =>
                     {
@@ -62,7 +62,7 @@ namespace OpenTibia.Game.CommandHandlers
                             return Promise.Break;
                         }
 
-                        byte afterIndex = afterContainer.GetIndex(command.Item);
+                        int afterIndex = afterContainer.GetIndex(command.Item);
 
                         if (beforeIndex != afterIndex)
                         {
