+++ mtanksl.OpenTibia.Game/Events/Tile/TileRefreshItemEventArgs.cs
@@ -4,7 +4,7 @@ namespace OpenTibia.Game.Events
 {
     public class TileRefreshItemEventArgs : GameEventArgs
     {
-        public TileRefreshItemEventArgs(Tile tile, Item item, byte index)
+        public TileRefreshItemEventArgs(Tile tile, Item item, int index)
         {
             Tile = tile;
 
@@ -17,6 +17,6 @@ namespace OpenTibia.Game.Events
 
         public Item Item { get; set; }
 
-        public byte Index { get; set; }
+        public int Index { get; set; }
     }
 }
